Separate hosted Nexus registration test
